pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autoppt"
version = "0.5.0"
authors = [
  { name="yeasy", email="yeasy@example.com" },
]
description = "Generate Professional Presentations in Seconds using AI"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
    "Topic :: Office/Business :: Financial :: Spreadsheet",
    "Topic :: Multimedia :: Graphics :: Presentation",
]
dependencies = [
    "anthropic>=0.49.0",
    "ddgs>=9.6.0",
    "google-genai>=1.8.0",
    "markitdown[pptx]>=0.1.0",
    "openai>=1.68.0",
    "Pillow>=10.0.0",
    "pydantic>=2.8.0",
    "python-dotenv>=1.0.0",
    "python-pptx>=0.6.23",
    "requests>=2.32.0",
    "streamlit>=1.42.0",
    "tqdm>=4.66.0",
    "trafilatura>=2.0.0",
    "wikipedia>=1.4.0",
]

[project.optional-dependencies]
dev = [
    "build>=1.2.2",
    "mypy>=1.14.0",
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
]

[project.urls]
"Homepage" = "https://github.com/yeasy/autoppt"
"Bug Tracker" = "https://github.com/yeasy/autoppt/issues"

[project.scripts]
autoppt = "autoppt.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["autoppt*"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true
